Move DataMapper setup and Seeder to helpers
